+++ src/convertQwertyToAlphabet/constants.ts
@@ -1,0 +1,57 @@
+/**
+ * qwerty 키보드 자판의 대소문자를 구분한 영어 알파벳을 한글 음소와 맵핑한 객체
+ */
+export const QWERTY_KEYBOARD_MAP = {
+  q: 'ㅂ',
+  Q: 'ㅃ',
+  w: 'ㅈ',
+  W: 'ㅉ',
+  e: 'ㄷ',
+  E: 'ㄸ',
+  r: 'ㄱ',
+  R: 'ㄲ',
+  t: 'ㅅ',
+  T: 'ㅆ',
+  y: 'ㅛ',
+  Y: 'ㅛ',
+  u: 'ㅕ',
+  U: 'ㅕ',
+  i: 'ㅑ',
+  I: 'ㅑ',
+  o: 'ㅐ',
+  O: 'ㅒ',
+  p: 'ㅔ',
+  P: 'ㅖ',
+  a: 'ㅁ',
+  A: 'ㅁ',
+  s: 'ㄴ',
+  S: 'ㄴ',
+  d: 'ㅇ',
+  D: 'ㅇ',
+  f: 'ㄹ',
+  F: 'ㄹ',
+  g: 'ㅎ',
+  G: 'ㅎ',
+  h: 'ㅗ',
+  H: 'ㅗ',
+  j: 'ㅓ',
+  J: 'ㅓ',
+  k: 'ㅏ',
+  K: 'ㅏ',
+  l: 'ㅣ',
+  L: 'ㅣ',
+  z: 'ㅋ',
+  Z: 'ㅋ',
+  x: 'ㅌ',
+  X: 'ㅌ',
+  c: 'ㅊ',
+  C: 'ㅊ',
+  v: 'ㅍ',
+  V: 'ㅍ',
+  b: 'ㅠ',
+  B: 'ㅠ',
+  n: 'ㅜ',
+  N: 'ㅜ',
+  m: 'ㅡ',
+  M: 'ㅡ',
+} as const;
