+++ src/disassembleCompleteCharacter/constants.ts
@@ -1,0 +1,1 @@
+export const NUMBER_OF_JUNGSEONG = 21;
