+++ src/romanize/constants.ts
@@ -1,0 +1,68 @@
+export const 중성_알파벳_발음 = {
+  // ------- 단모음
+  ㅏ: 'a',
+  ㅓ: 'eo',
+  ㅗ: 'o',
+  ㅜ: 'u',
+  ㅡ: 'eu',
+  ㅣ: 'i',
+  ㅐ: 'ae',
+  ㅔ: 'e',
+  ㅚ: 'oe',
+  ㅟ: 'wi',
+  // -------
+  // ------- 이중모음
+  ㅑ: 'ya',
+  ㅕ: 'yeo',
+  ㅛ: 'yo',
+  ㅠ: 'yu',
+  ㅒ: 'yae',
+  ㅖ: 'ye',
+  ㅘ: 'wa',
+  ㅙ: 'wae',
+  ㅝ: 'wo',
+  ㅞ: 'we',
+  ㅢ: 'ui',
+} as const;
+
+export const 초성_알파벳_발음 = {
+  // ------- 파열음
+  ㄱ: 'g',
+  ㄲ: 'kk',
+  ㅋ: 'k',
+  ㄷ: 'd',
+  ㄸ: 'tt',
+  ㅌ: 't',
+  ㅂ: 'b',
+  ㅃ: 'pp',
+  ㅍ: 'p',
+  // -------
+  // ------- 파찰음
+  ㅈ: 'j',
+  ㅉ: 'jj',
+  ㅊ: 'ch',
+  // -------
+  // ------- 마찰음
+  ㅅ: 's',
+  ㅆ: 'ss',
+  ㅎ: 'h',
+  // -------
+  // ------- 비음
+  ㄴ: 'n',
+  ㅁ: 'm',
+  ㅇ: '',
+  // -------
+  // ------- 유음
+  ㄹ: 'r',
+} as const;
+
+export const 종성_알파벳_발음 = {
+  ㄱ: 'k',
+  ㄴ: 'n',
+  ㄷ: 't',
+  ㄹ: 'l',
+  ㅁ: 'm',
+  ㅂ: 'p',
+  ㅇ: 'ng',
+  '': '',
+} as const;
