+++ src/susa/constants.ts
@@ -1,0 +1,29 @@
+export const SUSA_MAP = {
+  1: '하나',
+  2: '둘',
+  3: '셋',
+  4: '넷',
+  5: '다섯',
+  6: '여섯',
+  7: '일곱',
+  8: '여덟',
+  9: '아홉',
+  10: '열',
+  20: '스물',
+  30: '서른',
+  40: '마흔',
+  50: '쉰',
+  60: '예순',
+  70: '일흔',
+  80: '여든',
+  90: '아흔',
+  100: '백',
+} as const;
+
+export const SUSA_CLASSIFIER_MAP = {
+  1: '한',
+  2: '두',
+  3: '세',
+  4: '네',
+  20: '스무',
+} as const;
